==> RNAseq/scripts/01_check_names.Rmd <==
Checking how to reconcile kallisto file names and the excel key file

```{r}
library(tidyverse)
```

get the kallisto-generated file names.  These correspond to the file names from the sequencer.

```{r}
files <- dir(path = "../input/Kallisto_output/", include.dirs = TRUE)
kallisto_q <- str_extract(files, pattern = "q_[0-9]{3}(_d8)?") %>% 
  sort()

kallisto_q
```

read the key in
```{r}
key <- readxl::read_excel("../input/tube_no_legend_time_course_2018.xlsx")

key
```

Does the q_# in kallisto correspond to tube #?
```{r}
tube_no <- key %>% pull(tube_no) %>% 
  tolower() %>%
  str_replace("q_([1-9](_|$))", "q_00\\1") %>%
  str_replace("q_([1-9][0-9](_|$))", "q_0\\1")
tube_no

kallisto_q[! kallisto_q %in% tube_no]

tube_no[! tube_no %in% kallisto_q]
```
yes, all kallisto q #s are found in tube no, and almost all tube #s are found in kallisto
