==> RNAseq/scripts/03_Quick_DE_Analysis.Rmd <==
---
title: "03_Quick_DE_Analysis"
output: html_document
---

Quick analysis to calculate the number of DE genes per timepoint.

```{r}
library(tidyverse)
library(edgeR)
```


```{r}
load("../output/timecourseDGE.Rdata")
```

check the experimental factors so that we know what we have
```{r}
dge$samples$day_time <- str_c(dge$samples$sampling_day,"-",dge$samples$sampling_time)

cat("soil_trt: ", unique(dge$samples$soil_trt), "\n\n")
cat("genotype: ", unique(dge$samples$genotype), "\n\n")
cat("day_time: ", unique(dge$samples$day_time), "\n\n")
```

Create a tibble to hold data and models.
Each row of the tibble will correspond to one time point.
Each row has a subsetted dge that corresponds to that time point.
```{r}
models <- tibble(day_time=unique(dge$samples$day_time))
models <- models %>%
    separate(day_time, into=c("day", "time"), sep="-", remove=FALSE) %>%
  mutate(dge=map(day_time, ~ dge[,dge$samples$day_time==.]))

#Do NOT print the full models tibble to the screen, it slows everything way down!!
#models

models %>% select(-dge)
```

functions for calculating differential gene expression
```{r}
calcdisp <- function(dge) {
  design <- model.matrix(~soil_trt + block, data=dge$samples)
  dge <- estimateDisp(dge, design=design)
  dge
}

calcLRT <- function(dge) {
    design <- model.matrix(~soil_trt + block, data=dge$samples)
    fit <- glmFit(dge, design)
    lrt <- glmLRT(fit,coef = "soil_trtSBC_OLD")
    lrt
}

calcDEG <- function(lrt) {
  topTags(lrt, n=Inf, p.value = 0.05)
}
```

Calculate dispersion, LRT, and Toptags (p < 0.05) for each time point
```{r}
models <- models %>% mutate(dge=map(dge, calcdisp))
models <- models %>% mutate(lrt=map(dge, calcLRT))
models <- models %>% mutate(DEG0.05=map(lrt, calcDEG))
```

Save it.
```{r}
save(models, file="../output/TimeCourseDGE_LRT.Rdata")
```

Now count the number of differnetially expressed genes for each time point.  Also, fix day_time and day for better plotting.
```{r}
results <- models %>%
  mutate(n.DEG = map_int(DEG0.05, nrow)) %>%
  mutate(day_time = str_replace(day_time, "^([1-9]-)", "0\\1"),
         day=as.numeric(day)) %>%
  select(-dge, -lrt) %>%
  arrange(day_time)
```

plot it
```{r}
results %>%
  ggplot(aes(x=time, y=n.DEG)) +
  geom_col() +
  facet_grid(day ~ .)
ggsave("../output/time_course_DEGs.png", height=6, width=8)
```
