==> RNAseq/scripts/05_Iron.Rmd <==
---
title: "Iron Analysis"
output: html_notebook
---

Kazu has found a strong negative correlation between iron deficiency genes and response to live microves in exp1 and exp3.  Is this also true in the time course?

```{r}
library(tidyverse)
library(readxl)
library(edgeR)
```

get the iron data
```{r}
iron <- read_excel("../input/pnas.1916892116.sd01.xlsx", sheet = "iron deficiency reponsive genes") %>%
  select(GeneID=`Gene ID`, GeneSymbol=`Gene Symbol`, FC= `Fold Changes (-Fe/+Fe)`) %>%
  mutate(logFC=log2(FC))
iron
```

Now get the microbe response data
```{r}
load("../output/TimeCourseDGE_LRT.Rdata")
```

```{r}
names(models)
DEgenes <- models %>% 
  mutate(toptags = map(lrt, ~ topTags(.,n=Inf)), #pull the DE gene info out of each model
         toptags = map(toptags, ~ .$table), # get the table of DE info
         toptags = map(toptags, ~ rownames_to_column(., "transcript_ID"))) %>%
  unnest(toptags) %>%
  group_by(transcript_ID) %>%
  arrange(transcript_ID, FDR) %>% 
  select(day_time, transcript_ID, logFC, FDR)
DEgenes
```

Add Arabidopdis ID to each Brassica gene

First, get annotation
```{r}
annotation <- read_csv("../../Brapa_microbes/Annotation/output/v3.0annotation/Brapa_V3.0_annotated.csv")
annotation
```

now join the annotation
```{r}
DEgenes <- annotation %>%
  select(name, AGI) %>%
  right_join(DEgenes, by=c("name" = "transcript_ID"))
DEgenes
```

now add in the microbe data

```{r}
gene_compare <- DEgenes %>%
  inner_join(iron, by = c("AGI" = "GeneID"), suffix = c(".microbe", ".Fe"))
gene_compare
```

```{r}
gene_compare <- gene_compare %>%
  group_by(day_time) %>%
  nest()
gene_compare
```

```{r}
gene_compare <- gene_compare %>%
  mutate(cortest = map(data, ~ cor.test(.$logFC.microbe, .$logFC.Fe, method="spearman", exact=FALSE)))
```

```{r}
gene_compare %>% mutate(glance = map(cortest, broom::glance)) %>% unnest(glance) %>%
  arrange(p.value) %>% select(day_time, estimate, p.value) 
```
