==> RNAseq/scripts/06_two_factor_models.Rmd <==
---
title: "06 Two factor models"
output: html_document
---

```{r setup, include=FALSE}
#library(knitr)
knitr::opts_chunk$set(echo = TRUE)
#library(d3heatmap)
#library(flexdashboard)
library(edgeR);library(tidyverse)
library(readr);library(readxl)
library(stringr)

```


# loading DEG object created in "02_Normalize_DGE.Rmd"
```{r}
load("../output/timecourseDGE.Rdata")
```

check the experimental factors so that we know what we have
```{r}
dge$samples$day_time <- str_c(dge$samples$sampling_day,"-",dge$samples$sampling_time)

cat("soil_trt: ", unique(dge$samples$soil_trt), "\n\n")
cat("genotype: ", unique(dge$samples$genotype), "\n\n")
cat("day_time: ", unique(dge$samples$day_time), "\n\n")
```

Create a tibble to hold data and models.
Each row of the tibble will correspond to one time point.
Each row has a subsetted dge that corresponds to that time point.
```{r}
models <- tibble(day_time=unique(dge$samples$day_time))
models <- models %>%
  separate(day_time, into=c("day", "time"), sep="-", remove=FALSE) %>%
  mutate(dge=map(day_time, ~ dge[,dge$samples$day_time==.]))

#Do NOT print the full models tibble to the screen, it slows everything way down!!
#models

models %>% dplyr::select(-dge) #%>% View()
```

# MDS plot bcv (under construction)
```{r}
# MDS plot
mds <- plotMDS(dge,method = "bcv",labels=dge$samples$group,gene.selection = "pairwise",dim.plot = c(1,2),ndim=5)
# sample tibble
samples.tibble<-dge$samples %>% as_tibble() %>% bind_cols(data.frame(sample=row.names(dge$samples)),.)
# MDS plot data                                   
mds.pl <- as_tibble(mds$cmdscale.out) %>% 
  bind_cols(data.frame(sample=row.names(mds$cmdscale.out)),.) %>%
  inner_join(samples.tibble,by="sample")
# ggplot2 version
p.bcv<-mds.pl %>% ggplot(aes(x=V1,y=V2, shape=sampling_time, color=soil_trt,label=sampling_day)) + geom_point(size=3) + ggtitle("timecourse Samples DIM 1 vs 2 bcv")+ facet_grid(sampling_time~sampling_day) + theme_gray()
ggsave(p.bcv,filename="../output/timecourse.mds.bcv.1and2.png",width=12,height=8)

mds.pl %>% ggplot(aes(x=V1,y=V2, shape=sampling_time, color=soil_trt,label=sampling_day)) + geom_point(size=3) + ggtitle("Samples DIM 1 vs 2 bcv")+ facet_grid(.~block)
mds.pl %>% ggplot(aes(x=V2,y=V3, shape=sampling_time, color=soil_trt,label=sampling_day)) + geom_point(size=3) + ggtitle("Samples DIM 2 vs 3 bcv")+ facet_grid(.~block)
mds.pl %>% ggplot(aes(x=V3,y=V4, shape=sampling_time, color=soil_trt,label=sampling_day)) + geom_point(size=3) + ggtitle("Samples DIM 3 vs 4 bcv")+ facet_grid(.~block)
# no large block effects


```


# MDS plot logFC (under construction)
```{r}
# MDS plot
mds.logFC <- plotMDS(dge,method = "logFC",labels=dge$samples$group,gene.selection = "pairwise",dim.plot = c(1,2),ndim=5)
# sample tibble
samples.tibble<-dge$samples %>% as_tibble() %>% bind_cols(data.frame(sample=row.names(dge$samples)),.)
# MDS plot data                                   
mds.logFC.pl <- as_tibble(mds.logFC$cmdscale.out) %>% 
  bind_cols(data.frame(sample=row.names(mds$cmdscale.out)),.) %>%
  inner_join(samples.tibble,by="sample")
# ggplot2 version
p.logFC<-mds.logFC.pl %>% ggplot(aes(x=V1,y=V2, shape=sampling_time, color=soil_trt,label=sampling_day)) + geom_point(size=3) + ggtitle("timecourse Samples DIM 1 vs 2 logFC")+ facet_grid(sampling_time~sampling_day) + theme_gray()
ggsave(p.logFC,filename="../output/timecourse.mds.logFC.1and2.png",width=12,height=8)
# V2, V2
mds.logFC.pl %>% ggplot(aes(x=V2,y=V3, shape=sampling_time, color=soil_trt,label=sampling_day)) + geom_point(size=3) + ggtitle("timecourse Samples DIM 1 vs 2 logFC")+ facet_grid(sampling_time~sampling_day) + theme_gray()
ggsave(filename="../output/timecourse.mds.logFC.2and3.png",width=12,height=8)

# V3, V4
mds.logFC.pl %>% ggplot(aes(x=V3,y=V4, shape=sampling_time, color=soil_trt,label=sampling_day)) + geom_point(size=3) + ggtitle("timecourse Samples DIM 1 vs 2 logFC")+ facet_grid(sampling_time~sampling_day) + theme_gray()
ggsave(filename="../output/timecourse.mds.logFC.3and4.png",width=12,height=8)


mds.logFC.pl %>% ggplot(aes(x=V1,y=V2, shape=sampling_time, color=soil_trt,label=sampling_day)) + geom_point(size=3) + ggtitle("Samples DIM 1 vs 2 bcv")+ facet_grid(.~block)
mds.logFC.pl %>% ggplot(aes(x=V2,y=V3, shape=sampling_time, color=soil_trt,label=sampling_day)) + geom_point(size=3) + ggtitle("Samples DIM 2 vs 3 bcv")+ facet_grid(.~block)
mds.logFC.pl %>% ggplot(aes(x=V3,y=V4, shape=sampling_time, color=soil_trt,label=sampling_day)) + geom_point(size=3) + ggtitle("Samples DIM 3 vs 4 bcv")+ facet_grid(.~block)
# no large block effects


```



# "2-afternoon" data. y ~ trt + day + trt:day
```{r}
# subset data
dge.twoafternoon<-dge[,dge$samples$sampling_time=="2_afternoon"]
# reset normalization factor
dge.twoafternoon$samples$norm.factors<-1 # # No matter which case, non-normalize libraries.
# filtering
dge.twoafternoon<-dge.twoafternoon[rowSums(cpm(dge.twoafternoon)>1) >=3,keep.lib.sizes=FALSE]
# #relevel soil_trt to "SBC_OLD" 
dge.twoafternoon$samples<-dge.twoafternoon$samples %>% mutate(soil_trt=fct_relevel(soil_trt,"ATM_BLANK")) # %>% str() 
## model
dge.twoafternoon.design <- with(dge.twoafternoon$samples, model.matrix(~ soil_trt + sampling_day + soil_trt:sampling_day + block))
# estimateDisp
dge.twoafternoon <- estimateDisp(dge.twoafternoon,design = dge.twoafternoon.design)
## fit linear model
dge.twoafternoon.fit<- glmFit(dge.twoafternoon,dge.twoafternoon.design)
```

## get DEGs, soil_trt effects (coef=c("SBC_OLD"))
```{r}
twoafternoon.trtlive.lrt <- glmLRT(dge.twoafternoon.fit,coef = c("soil_trtSBC_OLD"))
## all
twoafternoon.trtlive.DEGs.all <- topTags(twoafternoon.trtlive.lrt,n = Inf,p.value = 1)$table
## pvalue < 0.05
twoafternoon.trtlive.DEGs.all %>% filter(FDR<0.05) %>% summarize(n()) # 1474
```

## get DEGs, soil_trt  interactions
```{r}
twoafternoon.interaction.trtlive.samplingday.lrt <- glmLRT(dge.twoafternoon.fit,coef = str_which(colnames(dge.twoafternoon.design),"soil_trtSBC_OLD:"))
## all
twoafternoon.interaction.trtlive.samplingday.lrt.DEGs.all <- topTags(twoafternoon.interaction.trtlive.samplingday.lrt,n = Inf,p.value = 1)$table
## pvalue < 0.05
twoafternoon.interaction.trtlive.samplingday.lrt.DEGs.all %>% filter(FDR<0.05) %>% summarize(n()) # 310
```

## get DEGs, any soil_trt effects combined 
```{r}
twoafternoon.any.trtlive.lrt <- glmLRT(dge.twoafternoon.fit,coef = str_which(colnames(dge.twoafternoon.design),"soil_trtSBC_OLD"))
## all
twoafternoon.any.trtlive.DEGs.all <- topTags(twoafternoon.any.trtlive.lrt,n = Inf,p.value = 1)$table
## pvalue < 0.05
twoafternoon.any.trtlive.DEGs.all %>% filter(FDR<0.05) %>% summarize(n()) # 2249
```

## get DEGs day effects only (). needs to be renamed (March 4, 2020)
* old: twoafternoon.any.trtlive
* new: twoafternoon.any.samplingday
```{r}
twoafternoon.any.samplingday.lrt <- glmLRT(dge.twoafternoon.fit,coef = str_which(colnames(dge.twoafternoon.design),"^sampling"))
## all
twoafternoon.any.samplingday.DEGs.all <- topTags(twoafternoon.any.samplingday.lrt,n = Inf,p.value = 1)$table
## pvalue < 0.05
twoafternoon.any.samplingday.DEGs.all %>% filter(FDR<0.05) %>% summarize(n()) # 15376
```

# diurnal data. One analysis for day 3+4 and a separate analysis 

## Days 3 and 4
```{r}
# subset data
dge.diurnal34<-dge[,dge$samples$sampling_day %in% c("03","04")]
dge.diurnal34$samples
# reset normalization factor
dge.diurnal34$samples$norm.factors<-1 # # No matter which case, non-normalize libraries.
# filtering
dge.diurnal34<-dge.diurnal34[rowSums(cpm(dge.diurnal34)>1) >=3,keep.lib.sizes=FALSE]
# relevel soil_trt to "SBC_OLD" 
dge.diurnal34$samples<-dge.diurnal34$samples %>% mutate(soil_trt=fct_relevel(soil_trt,"ATM_BLANK")) # %>% str() 
# relevel sampling_time to "1_morn"
dge.diurnal34$samples<-dge.diurnal34$samples %>% mutate(sampling_time=fct_relevel(sampling_time,"1_morn")) # %>% str() 
# model
dge.diurnal34.design <- with(dge.diurnal34$samples, model.matrix(~ soil_trt + sampling_time + soil_trt:sampling_time + block))
colnames(dge.diurnal34.design)
# estimateDisp
dge.diurnal34 <- estimateDisp(dge.diurnal34,design = dge.diurnal34.design)
## fit linear model
dge.diurnal34.fit <- glmFit(dge.diurnal34,dge.diurnal34.design)
```


trt X time interaction effects
```{r}
# get DEGs, interaction effects at any time points (under construction)
dge.diurnal34.interaction.trtlive.time.lrt <- glmLRT(dge.diurnal34.fit,coef = str_which(colnames(dge.diurnal34.design),":")) 
dge.diurnal34.interaction.trtlive.time.DEGs.all <- topTags(dge.diurnal34.interaction.trtlive.time.lrt,n = Inf,p.value = 1)$table

## pvalue < 0.05
dge.diurnal34.interaction.trtlive.time.DEGs <- topTags(dge.diurnal34.interaction.trtlive.time.lrt,n = Inf,p.value = 0.05)$table
nrow(dge.diurnal34.interaction.trtlive.time.DEGs) # 4???? something is wrong???
```
BraA05g007030.3C: AT2G38640.1 Unknown
BraA02g000300.3C: AT5G01210 HXXXD-type acyl-transferase family protein
BraA06g008660: AT1G13110: cytochrome P450, family 71 subfamily B, polypeptide 7 
(plastid located)
BraA10g017550: AT5G59720: HeatShock Protein


trt only

```{r}
# get DEGs, trt
dge.diurnal34.trtlive.lrt <- glmLRT(dge.diurnal34.fit,coef = str_which(colnames(dge.diurnal34.design),"^soil_trtSBC_OLD$")) ## all
dge.diurnal34.trtlive.DEGs.all <- topTags(dge.diurnal34.trtlive.lrt,n = Inf,p.value = 1)$table
## pvalue < 0.05
dge.diurnal34.trtlive.DEGs <- topTags(dge.diurnal34.trtlive.lrt,n = Inf,p.value = 0.05)$table;nrow(dge.diurnal34.trtlive.DEGs) # 4593
```

What about trt + trt:interaction
```{r}
# get DEGs, trt
dge.diurnal34.any.trtlive.lrt <- glmLRT(dge.diurnal34.fit,coef = str_which(colnames(dge.diurnal34.design),"soil_trtSBC_OLD")) ## all
dge.diurnal34.any.trtlive.DEGs.all <- topTags(dge.diurnal34.any.trtlive.lrt, n = Inf,p.value = 1)$table
## pvalue < 0.05
dge.diurnal34.any.trtlive.DEGs <- topTags(dge.diurnal34.any.trtlive.lrt,n = Inf,p.value = 0.05)$table;nrow(dge.diurnal34.any.trtlive.DEGs) # 4593
```

dge.diurnal34.any.trtlive.DEGs.all

time only

```{r}
# get DEGs, time
dge.diurnal34.time.lrt <- glmLRT(dge.diurnal34.fit,coef = str_which(colnames(dge.diurnal34.design),"^sampling")) ## all
dge.diurnal34.time.DEGs.all <- topTags(dge.diurnal34.time.lrt,n = Inf,p.value = 1)$table
## pvalue < 0.05
dge.diurnal34.time.DEGs <- topTags(dge.diurnal34.time.lrt,n = Inf,p.value = 0.05)$table;nrow(dge.diurnal34.time.DEGs) # 6892
```

## Days 13 and 14

```{r}
# subset data
dge.diurnal1314<-dge[,dge$samples$sampling_day %in% c("13","14")]
dge.diurnal1314$samples
# reset normalization factor
dge.diurnal1314$samples$norm.factors<-1 # # No matter which case, non-normalize libraries.
# filtering
dge.diurnal1314<-dge.diurnal1314[rowSums(cpm(dge.diurnal1314)>1) >=3,keep.lib.sizes=FALSE]
# relevel soil_trt to "SBC_OLD" 
dge.diurnal1314$samples<-dge.diurnal1314$samples %>% mutate(soil_trt=fct_relevel(soil_trt,"ATM_BLANK")) # %>% str() 
# relevel sampling_time to "1_morn"
dge.diurnal1314$samples<-dge.diurnal1314$samples %>% mutate(sampling_time=fct_relevel(sampling_time,"1_morn")) # %>% str() 
# model
dge.diurnal1314.design <- with(dge.diurnal1314$samples, model.matrix(~ soil_trt + sampling_time + soil_trt:sampling_time + block))
colnames(dge.diurnal1314.design)
# estimateDisp
dge.diurnal1314 <- estimateDisp(dge.diurnal1314,design = dge.diurnal1314.design)
## fit linear model
dge.diurnal1314.fit <- glmFit(dge.diurnal1314,dge.diurnal1314.design)
```

trt:time interaction
```{r}
# get DEGs, interaction effects 
dge.diurnal1314.interaction.trtlive.time.lrt <- glmLRT(dge.diurnal1314.fit,coef = str_which(colnames(dge.diurnal1314.design),":")) 
dge.diurnal1314.interaction.trtlive.time.DEGs.all <- topTags(dge.diurnal1314.interaction.trtlive.time.lrt,n = Inf,p.value = 1)$table

## pvalue < 0.05
dge.diurnal1314.interaction.trtlive.time.DEGs <- topTags(dge.diurnal1314.interaction.trtlive.time.lrt,n = Inf,p.value = 0.05)$table
nrow(dge.diurnal1314.interaction.trtlive.time.DEGs) # 1???? something is wrong???

```

trt only
```{r}
# get DEGs, trt
dge.diurnal1314.trtlive.lrt <- glmLRT(dge.diurnal1314.fit,coef = str_which(colnames(dge.diurnal1314.design),"soil_trtSBC_OLD$")) ## all
dge.diurnal1314.trtlive.DEGs.all <- topTags(dge.diurnal1314.trtlive.lrt,n = Inf,p.value = 1)$table

## pvalue < 0.05
dge.diurnal1314.trtlive.DEGs <- topTags(dge.diurnal1314.trtlive.lrt,n = Inf,p.value = 0.05)$table
nrow(dge.diurnal1314.trtlive.DEGs) # 26
```

What about trt + trt:time?
```{r}
# get DEGs, trt
dge.diurnal1314.any.trtlive.lrt <- glmLRT(dge.diurnal1314.fit,coef = str_which(colnames(dge.diurnal1314.design),"soil_trtSBC_OLD")) ## all
dge.diurnal1314.any.trtlive.DEGs.all <- topTags(dge.diurnal1314.any.trtlive.lrt,n = Inf,p.value = 1)$table
## pvalue < 0.05
dge.diurnal1314.any.trtlive.DEGs <- topTags(dge.diurnal1314.any.trtlive.lrt,n = Inf,p.value = 0.05)$table;nrow(dge.diurnal1314.any.trtlive.DEGs) # 422
```

Above asks for the interaction effects...what about time on its own?

```{r}
# get DEGs, time
dge.diurnal1314.time.lrt <- glmLRT(dge.diurnal1314.fit,coef = str_which(colnames(dge.diurnal1314.design),"^sampling")) ## all
dge.diurnal1314.time.DEGs.all <- topTags(dge.diurnal1314.time.lrt,n = Inf,p.value = 1)$table
## pvalue < 0.05
dge.diurnal1314.time.DEGs <- topTags(dge.diurnal1314.time.lrt,n = Inf,p.value = 0.05)$table;nrow(dge.diurnal1314.time.DEGs) # 
```



# Adding annotation 
## prep
```{r}
# annotation file for v3.0annotation
Br.v3.0.At.BLAST <- read_csv(file.path("..","Annotation_copy","output","v3.0annotation","Brapa_v3.0_annotated.csv")) 
# This annotation is redundant with name (Br grene). Eg 
Br.v3.0.At.BLAST %>% filter(name=="BraA01g040570.3C")
# reduce the redundancy (112418)
Br.v3.0anno.At.BLAST.highscore <- Br.v3.0.At.BLAST %>% group_by(name) %>% arrange(desc(score)) %>% dplyr::slice(1)
# function for adding annotation
## get object name https://stackoverflow.com/questions/14577412/how-to-convert-variable-object-name-into-string
myfunc <- function(v1) {
  deparse(substitute(v1))
}
myfunc(foo)
# adding annotation and write_csv adding ".v3.0anno.csv" to the object name.
addAnno <- function(DGE) {
  temp <- DGE %>% 
    rownames_to_column(var="genes") %>% 
    left_join(Br.v3.0anno.At.BLAST.highscore,by=c(genes="name")) %>% 
    dplyr::select(genes,names(DGE),AGI, At_symbol, At_short_description, perc_ID)
  print(deparse(substitute(DGE)))
  write_csv(temp, path=file.path("..","output",paste(deparse(substitute(DGE)),".v3.0anno.csv",sep="")))
  return(temp)
} 
```

## add annotation with full name objects (all) and write it in csv (stored in "output" directory)
```{r}
# two afternoon trt
twoafternoon.trtlive.DEGs.all.v3.0anno <- addAnno(twoafternoon.trtlive.DEGs.all)

# two afternoon interaction
twoafternoon.interaction.trtlive.samplingday.lrt.DEGs.all.v3.0anno <- addAnno(twoafternoon.interaction.trtlive.samplingday.lrt.DEGs.all)

# two afternoon trt all
twoafternoon.any.trtlive.DEGs.all.v3.0anno <- addAnno(twoafternoon.any.trtlive.DEGs.all)

# diel interaction
dge.diurnal34.interaction.trtlive.time.DEGs.all.v3.0anno <- addAnno(dge.diurnal34.interaction.trtlive.time.DEGs.all)
dge.diurnal1314.interaction.trtlive.time.DEGs.all.v3.0anno<-addAnno(dge.diurnal1314.interaction.trtlive.time.DEGs.all)

# diel trt
dge.diurnal34.trtlive.DEGs.all.v3.0anno <- addAnno(dge.diurnal34.trtlive.DEGs)
dge.diurnal1314.trtlive.DEGs.all.v3.0anno <- addAnno(dge.diurnal1314.trtlive.DEGs.all)

# diel trt all
dge.diurnal34.any.trtlive.DEGs.all.v3.0anno <- addAnno(dge.diurnal34.any.trtlive.DEGs.all)
dge.diurnal1314.any.trtlive.DEGs.all.v3.0anno <- addAnno(dge.diurnal1314.any.trtlive.DEGs.all)

```

# Session info
```{r}
sessionInfo()
```
