==> RNAseq/scripts/10_plotting.Rmd <==
---
title: "plot time course"
output: html_notebook
---


```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE,error=TRUE)
library(edgeR);library(tidyverse)
```

```{r}
load("../output/timecourseDGE.Rdata")
dge.diurnal34 <- dge[,dge$samples$sampling_day %in% c("03","04")]
```

```{r}
cpm34 <- cpm(dge.diurnal34, log=TRUE)
colnames(cpm34) <- dge.diurnal34$sample$group
cpm34 <- cpm34 %>%
  as.data.frame() %>%
  rownames_to_column(var="gene_ID") %>%
  rename_all(str_remove,"FPsc-") %>%
  rename_all(str_remove,"FPsc-ATM_") %>%
  gather(key="sample", value="log2cpm", -gene_ID) %>%
  mutate(sample=str_remove(sample, "\\..*"),
         trt=str_extract(sample, "BLANK|SBC_OLD"),
         #time=str_remove(sample, ".*(04|03)-")
         time=str_remove(sample, "(SBC_OLD-|ATM_BLANK-)")
         )
unique(cpm34$time)
```

```{r}
plotgene <- function(cpm=cpm34, target="BraA05g007030.3C") {
  cpm %>% 
    filter(gene_ID %in% target) %>%
    ggplot(aes(y=log2cpm, x=time, color=trt, group=trt)) +
    geom_smooth(span=.2) +
    geom_point() +
    facet_wrap(~gene_ID, ncol=2, scales = "free_y") +
    theme(axis.text.x = element_text(angle = 90, hjust=1))
}

plotgene()

interaction.genes <- c("BraA05g007030.3C", "BraA02g000300.3C", "BraA06g008660.3C",	 "BraA10g017550.3C")

plotgene(target=interaction.genes)
```
