==> RNAseq/scripts/11_VarianceTimeCourse.Rmd <==
---
title: "Variance Time Course"
output: html_notebook
---


```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE,error=TRUE)
library(edgeR);library(tidyverse)
```

```{r}
load("../output/timecourseDGE.Rdata")
dge.2 <- dge[,dge$samples$sampling_time %in% c("2_afternoon")]
```

```{r}
cpm.2 <- cpm(dge.2, log=TRUE)
colnames(cpm.2) <- str_c(dge.2$samples$soil_trt, ".", dge.2$samples$sampling_day)
head(cpm.2[,1:10])
```

```{r}
cpm.2 <- cpm.2 %>%
  na.omit() %>%
  as.data.frame() %>%
  rownames_to_column("gene_ID") %>%
  gather(key=sample, value=logcpm, -gene_ID) %>% 
  mutate(trt=str_extract(sample,"(ATM_BLANK|SBC_OLD)"),
         day=str_extract(sample, "[0-9]{2}"))
cpm.2
```

filter to remove low-expressed genes
```{r}
cpm.2.filter  <- cpm.2 %>%
  group_by(gene_ID) %>%
  mutate(mean_exp_for_gene=mean(logcpm)) %>%
  filter(mean_exp_for_gene >= 1) #2 cpm
cpm.2.filter %>% arrange(gene_ID)
```


calculate cv and keep top 1000

```{r}
cv <- cpm.2.filter %>%
  summarize(cv = mean(logcpm)/sd(logcpm)) %>%
  arrange(desc(cv))



cpm.1000 <- cv[1:1000,] %>% left_join(cpm.2.filter, by="gene_ID")

cpm.1000
```


```{r}
cpm.summary_gene <- cpm.1000 %>% group_by(gene_ID, day, trt) %>%
  summarize(cv=mean(logcpm)/sd(logcpm)) #recalculate because now want to separate by trt
cpm.summary_gene
```

```{r}
cpm.summary_day <- cpm.summary_gene %>%
  group_by(day, trt) %>%
  summarise(mean_cv=mean(cv,na.rm=TRUE), sem=sd(cv,na.rm = TRUE)/sqrt(n()))

cpm.summary_day
  
```

```{r}
cpm.summary_day %>%
  ggplot(aes(x=day, group=trt, y=mean_cv, color=trt)) +
  geom_line()
```
