==> RNAseq/scripts/12_GSEA_custom_category.Rmd <==
---
title: "12_GSEA_custom_category"
author: "Kazu"
date: "3/25/2021"
output: html_document
editor_options: 
  chunk_output_type: console
---

* History
** 

* To Do
**
```{r setup, include=FALSE}
rm(list=ls())
knitr::opts_chunk$set(error = TRUE,warning=FALSE)
library(tidyverse);library(readr);library(readxl);library(openxlsx)
# The following setting is important for WGCNA, do not omit.
 options(stringsAsFactors = FALSE)
# install.packages(c("dynamicTreeCut", "cluster", "flashClust", "Hmisc", "reshape", "foreach", "doParallel") ) 
library(WGCNA);library(flashClust) # for WGCNA in Whitney
#allowWGCNAThreads(nThreads = NULL) # for Whitney
#enableWGCNAThreads(4) # for Whitney
library(scales) # for muted function
library(ggdendro) # for dendrogram
library(cowplot)
library(ggnewscale)

```


## 1) You could do GSEA separately for live vs dead for each time point. You would start with the script https://github.com/MaloofLab/Brapa-microbes-timecourse-2018/blob/master/scripts/02_Normalize_DGE.Rmd to get those genes. (Julin's comment 032621)

# CLS: Categorical (e.g tumor vs normal) class file format (*.cls) (http://software.broadinstitute.org/cancer/software/gsea/wiki/index.php/Data_formats#CLS:_Categorical_.28e.g_tumor_vs_normal.29_class_file_format_.28.2A.cls.29)
```{r}
# for log2cpmSample
load(file="../output/timecourseDGE.Rdata")
dge$samples %>% split(.$sampling_time)
# 
dge$samples %>% dplyr::count(sampling_day,sampling_time)

dge$samples %>% dplyr::count(sampling_day,sampling_time,soil_trt)
# 
test.map <- dge$samples %>% split(.$sampling_time) %>% map(~split(., .$sampling_day)) 
```

# divide expression data & csl file (phenotype) for GSEA
```{r}
log2cpmSample <- read_tsv(file.path("..","output","log2cpmSample.txt.gz"))
for(i in 1:5) {
  for(n in 1:length(names(test.map[[i]]))) {
    # sampleID.temp <- dge$samples %>% filter(sampling_time==names(test.map)[i],sampling_day==names(test.map[[i]])[n]) %>% rownames_to_column(var="sampleID") %>% dplyr::select(sampleID) %>% as_vector()
    x <- log2cpmSample %>% dplyr::select(gene,matches(names(test.map)[i]) & matches(names(test.map[[i]])[n]))# %>% colnames()
    # write cls
    Total <- length(colnames(x)[-1]);print(Total)
description <- "# LIVE BLANK";print(description) #  ("# (space) (class 0 name) (space) (class 1 name)")
Live.Blank <- ifelse(str_detect(colnames(x)[-1],"SBC_OLD"),0,1) # "live" is zero 
print(Live.Blank)
# writing cls file
cat(paste(Total,"2","1","\n"),file=file.path("..","output",paste(names(test.map)[i],names(test.map[[i]])[n],"d","cls",sep=".")))
cat(paste(description,"\n"),file=file.path("..","output",paste(names(test.map)[i],names(test.map[[i]])[n],"d","cls",sep=".")),append=TRUE)
cat(Live.Blank,file=file.path("..","output",paste(names(test.map)[i],names(test.map[[i]])[n],"d","cls",sep=".")),append=TRUE)
# writing expression data
write_tsv(x,file.path("..","output",paste(names(test.map)[i],names(test.map[[i]])[n],"d","txt",sep=".")))
  }
}
```

# 
```{r}
# custom categories list
load(file.path("..","Annotation_copy","output","v3.0annotation","GSEA_cust_rootshoot_updown2.list.Rdata"))
names(GSEA_cust_rootshoot_updown2.list)
GSEA_cust_rootshoot_updown2.list[["BraA01g000020.3C"]]
```

```{r}
read.gsea_report_files_timecourse_custom <- function(gsea_report_path,GSEA.FDR.cutoff) {
# open xsl files and save as csvs
# gsea_report
gsea_report_files <- list.files(path=gsea_report_path,full.names=FALSE,include.dirs = FALSE,recursive=TRUE, pattern="(gsea_report_for)([[:print:]]+)(\\.xls)")
gsea_report_files
# gsea_report (full name)
gsea_report_files2 <- list.files(path=gsea_report_path,full.names=TRUE,recursive=TRUE, pattern="(gsea_report_for)([[:print:]]+)(\\.xls)")
# read xls file (file name is "xls", actually tsv. That is why read_xls did not work)
#GSEA_reports<-lapply(gsea_report_files2, function(x) read_tsv(file=file.path(x)))
GSEA_reports <- gsea_report_files2 %>% map(~read_tsv(.))
# name
names(GSEA_reports) <- gsea_report_files
  gsub(".xls","",gsub("([[:print:]]+)(/)([[:print:]]+)","\\3",gsea_report_files))
# convert list into one data.frame
GSEA_reports.DF <- GSEA_reports %>% enframe(name="report") %>% unnest(value)
#GSEA_reports.DF %>% View()
# filter only significnt `FDR q-val` (GSEA.FDR.cutoff)
GSEA_reports.DF <- GSEA_reports.DF %>% filter(`FDR q-val`< GSEA.FDR.cutoff)
# 
GSEA_reports.DF <- GSEA_reports.DF %>% separate(report,into=c("report2","report3"),sep="/")  
#
#GSEA_reports.DF$LorD <- gsub("(gsea_report_for_)(0|1)([[:print:]]+)","\\2",GSEA_reports.DF$report3)
#GSEA_reports.DF <- GSEA_reports.DF %>% mutate(LorD=ifelse(LorD==0,"Live","Dead"))
GSEA_reports.DF %>% View()
ifelse(dim(GSEA_reports.DF)[1]==0,print("no significnat categories"),return(GSEA_reports.DF))
}
```

```{r}
gsea_report_path=file.path("~","Box","gsea_home","output","mar26")
GSEA_reports.DF <- read.gsea_report_files_timecourse_custom(gsea_report_path,GSEA.FDR.cutoff=1) # read all files
# separate sampling_time, sampling_day, soil_trt
GSEA_reports.DF %>% View()

```

# select custom categories GSEA and format
```{r}
GSEA_reports.DF.custom <- GSEA_reports.DF %>% filter(str_detect(report2,"custom"))
# separate time and day
GSEA_reports.DF.custom$sampling_time <- gsub("(1_morn|2_afternoon|3_evening_5.30|4_night_1|5_night_2)(\\.)(01|02|03|04|06|08|10|13|14)(.custom.Gsea.)([[:print:]]+)","\\1", GSEA_reports.DF.custom$report2)
GSEA_reports.DF.custom$sampling_day <- gsub("(1_morn|2_afternoon|3_evening_5.30|4_night_1|5_night_2)(\\.)(01|02|03|04|06|08|10|13|14)(.custom.Gsea.)([[:print:]]+)","\\3", GSEA_reports.DF.custom$report2)
# Live or Blank
GSEA_reports.DF.custom <- GSEA_reports.DF.custom %>% mutate(LorB=ifelse(str_detect(report3,"_LIVE_"),"L","B"))
# check
GSEA_reports.DF.custom %>% View()

```

#############
### plot GSEA results
###########################

# calculate numbers of genes in each 
```{r}
GSEA_cust <- read_tsv(file.path("..","Annotation_copy","output","v3.0annotation","filtered.custom.data.Total.rootshoot.v3.gmx")) %>% dplyr::slice(-1) # # remove row1 that has "na"
genes.count <- GSEA_cust %>% 
  pivot_longer(cols=colnames(GSEA_cust)) %>% 
  #mutate(name=str_to_upper(name)) %>%
  group_by(name) %>% drop_na() %>% summarise(n=n()) %>% 
  dplyr::rename(NAME=name) 
# fix NAME
genes.count2 <- genes.count %>% mutate(NAME=str_to_upper(NAME)) %>% 
  mutate(NAME=str_replace_all(NAME,"KAILASAM2019", "KAILASAM_2019")) %>%
  mutate(NAME=str_replace_all(NAME,"MINUSPI_1D","MINUSPI-1D")) %>%
  mutate(NAME=str_replace_all(NAME,"MINUSPI_3D","MINUSPI-3D")) %>%
  mutate(NAME=str_replace_all(NAME,"AARABI2016","AARABI_2016")) %>%
  mutate(NAME=str_replace_all(NAME,"KIM2019","KIM_2019")) %>%
  mutate(NAME=str_replace_all(NAME,"WANG2003","WANG_2003")) %>%
   mutate(NAME=str_replace_all(NAME,"COLDKILIAN","COLD_KILIAN")) %>%
  mutate(NAME=str_replace_all(NAME,"BL_SEEDLINGS","plusBL_Nemhauser_2006_seedlings")) %>%
  mutate(NAME=str_replace_all(NAME,"IAA_SEEDLINGS","plusIAA_Nemhauser_2006_seedlings")) #%>%
#    mutate(NAME=str_remove(NAME,"_FC"))

```

# revised for day_time_LorB 
```{r}
sampling_day_time.summary <- dge$samples %>% dplyr::count(sampling_day,sampling_time)

sampling_day_time_soil.summary <- dge$samples %>% dplyr::count(sampling_day,sampling_time,soil_trt) %>% mutate(LorB=ifelse(soil_trt=="ATM_BLANK","B","L")) %>% unite(time_day_LorB,sampling_time,sampling_day,LorB,remove=FALSE) %>% unite(day_time_LorB,sampling_day,sampling_time,LorB,remove=FALSE)

# "all"  
all <- tibble(sampling_day_time_soil = rep(sampling_day_time_soil.summary$day_time_LorB,each=length(genes.count2$NAME)),
              NAME=rep(genes.count2$NAME,dim(sampling_day_time_soil.summary)[1])) %>% unite(day_time_LorB_NAME,sampling_day_time_soil,NAME,remove=FALSE) 

```

### modify following scripts ####
# modify GSEA plot function
```{r}
# plot function
GSEA_reports.DF.plot.nodendrogram.custom.notitle <- function(data = GSEA_reports.DF.custom,
                                            GSEA.FDR.cutoff = cutoff,
                                            top.margin=830,
                                            bottom.margin=40,
                                            legend.height=20,
                                            legend.width=25,
                                            LorB=c("L","B")) {
# modify this for timecourse_custom
GSEA_reports.DF.formatted <- data  %>%  dplyr::select(sampling_time, sampling_day,LorB,NAME,`FDR q-val`) %>% mutate(`FDR q-val`=as.numeric(str_replace(`FDR q-val`,"^0$","0.01"))) %>% mutate(`-log10 FDR q-val`=-log10(`FDR q-val`)) %>% dplyr::select(-`FDR q-val`) %>% unite(day_time_LorB,sampling_day,sampling_time, LorB,remove=TRUE) %>% arrange(NAME)
## check
print("GSEA_reports.DF.formatted")
GSEA_reports.DF.formatted %>% View()
# replace "NA" to 0
GSEA_reports.DF.formatted[is.na(GSEA_reports.DF.formatted)] <- 0
# remove all colums are zero (otherwise resulting "NaN" after scale())

### there are bugs to fill in missing NAME and missing time_day_LorB below
GSEA_reports.DF.formatted <- GSEA_reports.DF.formatted %>% unite(day_time_LorB_NAME, day_time_LorB,NAME,remove=FALSE) 
missing <- all %>% anti_join(GSEA_reports.DF.formatted,by="day_time_LorB_NAME") %>% mutate(`-log10 FDR q-val`=0) %>% rename(day_time_LorB=sampling_day_time_soil)
print("missing");print(missing)
GSEA_reports.DF.formatted.all <- GSEA_reports.DF.formatted %>% bind_rows(missing)
  
## relevel
 #  GSEA_reports.DF.formatted2 <- GSEA_reports.DF.formatted2 %>% 
 # mutate(NAME=factor(NAME,levels=c(as.vector(dendro.y.NAME$NAME),as.vector(missing.categories$NAME)))) 
print("GSEA_reports.DF.formatted")
#GSEA_reports.DF.formatted.all %>% View()
# LorB
if(LorB=="L") {
  GSEA_reports.DF.formatted.all <- GSEA_reports.DF.formatted.all %>% filter(str_detect(day_time_LorB,"_L"))
} else {
    GSEA_reports.DF.formatted.all <- GSEA_reports.DF.formatted.all %>% filter(str_detect(day_time_LorB,"_B"))
    }
print("after LorB selection")
GSEA_reports.DF.formatted.all %>% View()
# simplify NAME=x-axis
GSEA_reports.DF.formatted.all <- GSEA_reports.DF.formatted.all %>% mutate(category2=
  str_to_upper(NAME) %>%
  str_replace_all("KAILASAM2019", "KAILASAM_2019") %>%
 str_replace_all("MINUSPI_1D","MINUSPI-1D") %>%
  str_replace_all("MINUSPI_3D","MINUSPI-3D") %>%
  str_replace_all("AARABI2016","AARABI_2016") %>%
  str_replace_all("KIM2019","KIM_2019") %>%
  str_replace_all("WANG2003","WANG_2003") %>%
   str_replace_all("COLDKILIAN","COLD_KILIAN") %>%
  str_replace_all("BL_SEEDLINGS","plusBL_Nemhauser_2006_seedlings") %>%
  str_replace_all("IAA_SEEDLINGS","plusIAA_Nemhauser_2006_seedlings") ) %>% 
  separate(category2,into=c("category3","au","yr","tissue3","FC","up_down"),sep="_") %>% # View()
  mutate(category3=category3 %>% 
           str_replace_all("MINUSMG","MG-") %>%
           str_replace_all("PLUSMG","MG+") %>%
           str_replace_all("MINUSFE","FE-") %>%
    str_replace_all("PLUSFE","FE_PLUS") %>%
  str_replace_all("MINUSPI-1D","PI-1D-") %>%
    str_replace_all("PLUSPI-1D","PI-1D+") %>%
  str_replace_all("MINUSPI-3D","PI-3D+") %>%
    str_replace_all("PLUSPI-3D","PI-3D+") %>%
  str_replace_all("MINUSN","N-") %>%
  str_replace_all("PLUSN","N+") %>%
  str_replace_all("MINUSMN","MN-") %>%
  str_replace_all("PLUSMN","MN+") %>% 
  str_replace_all("MINUSP","P-") %>% 
    str_replace_all("PLUSP","P+") %>% 
    str_replace_all("PLUSIAA","IAA+") %>% 
    str_replace_all("PLUSAL","AL+") %>%
    str_replace_all("MINUSCA","CA-") %>% 
      str_replace_all("PLUSCA","CA+") %>%
    str_replace_all("MINUSK","K-") %>% 
    str_replace_all("PLUSK","K+") %>% 
    str_replace_all("MINUSS","S-") %>% 
    str_replace_all("PLUSS","S+")) %>%
  drop_na(category3) %>%
  unite(category3.au,c("category3","au"),remove=FALSE) %>% 
  mutate(au2 = str_replace_all(au,"KAILASAM","(1)") %>% 
           str_replace_all("KIM","(2)") %>% 
           str_replace_all("NISHIDA","(3)") %>%
           str_replace_all("NIU","(4)") %>%
           str_replace_all("KILIAN","(5)") %>%
           str_replace_all("RODRIGUEZ-CELMA","(6)") %>%
           str_replace_all("LIU","(7)") %>%
           str_replace_all("PENG","(8)") %>%
           str_replace_all("AARABI","(9)") %>%
           str_replace_all("LIGABA-OSENA","(10)") %>%
           str_replace_all("NEMHAUSER","(11)") %>%
           str_replace_all("WANG","(12)")) %>%
  unite(category3.tissue.au.up_down,c("category3","tissue3","up_down","au2"),sep=" ",remove=FALSE) 
# print("after simplify category name)
GSEA_reports.DF.formatted.all %>% View()


# plot (base)
GSEA_reports.DF.plot <- GSEA_reports.DF.formatted.all %>% 
  ggplot(aes(x=category3.tissue.au.up_down,y=day_time_LorB)) + geom_tile(aes(fill=`-log10 FDR q-val`),colour="black")  + scale_fill_gradient2(limit=c(0,3), high=muted("magenta")) 

# + scale_x_discrete(sec.axis=dup_axis())# not working, ,vjust=-0.0001
GSEA_reports.DF.plot <- GSEA_reports.DF.plot +     theme(axis.text.x=element_text(size=9,angle=90,vjust=0.7),
      axis.text.y=element_text(size=10),
      axis.title=element_text(size=10),
      axis.ticks = element_blank(),
      strip.text.y = element_text(angle=0),
      panel.background = element_rect(fill = "white",colour="black"),
      plot.title=element_text(size=20),
      axis.line=element_blank()) + 
labs(x="",y="",fill="-log10\n FDR")

print(GSEA_reports.DF.plot)
##  gene number, custom category, etc. shape is growth increase/decrease (instead of UP/DOWN), plus correlation of modules to leaf_avg_std from modTrait.leaf_avg_std.RP2.cor.summary.voom5
# check genes.count2
print("genes.count2 after modules if/else")
print(genes.count2)
str(genes.count2)
print("dim(genes.count2)")
print(dim(genes.count2))
# genes.count2 <- genes.count2 %>% mutate(y=seq(dim(genes.count2)[1],1)) # bottom is y=1, 
# draw gene number inside heatmap
  #   GSEA_reports.DF.plot.alt <- GSEA_reports.DF.plot + 
  #     # for tissue for modules
  #     geom_point(aes(x=-0.35,y=y,color=tissue2),stroke=3, size=.8,inherit.aes = FALSE, data=genes.count2) +
  #     scale_color_manual(name = "", values = c("root"="brown", "leaf"="lawngreen")) + 
  #     #scale_shape_manual(values=c("increase"=2,"decrease"=6)) +
  #     labs(shape="") + #scale_shape_identity(name="") + 
  #     # for gene number (="n")
  # geom_point(aes(x=0,y=y,size=n),color = "red",  inherit.aes = FALSE, data=genes.count2) +
  #     new_scale_fill() + # error
  #     geom_point(aes(x=0.35,y=y,fill=Pct_Var_Explained,shape=Effect), size=3,inherit.aes = FALSE, data=genes.count2)  +
  #     scale_fill_gradient(limit=c(0,10)) +
  #     scale_shape_manual(values=c("Increase"=24,"Decrease"=25)) + # trial to heve different shape 
  #   coord_cartesian(xlim = c(0,4)) 
  #   GSEA_reports.DF.plot.alt # needs to work on abs(FC.mean) scale
### Adjust legend.
GSEA_reports.DF.plot.alt <- GSEA_reports.DF.plot + labs(title=NULL) + theme(
  legend.position = "top",legend.title=element_text(size=10),legend.text=element_text(size=10,angle=90),legend.key.height=unit(legend.height,"pt"),legend.key.width = unit(legend.width,"pt"),# Top, right, bottom, left
  plot.margin = margin(top.margin, 0, bottom.margin, 0, "pt")) # Top, right, bottom, left.
return(GSEA_reports.DF.plot.alt)
}
```
# trial
```{r}
# Live
GSEA.timecourse_soil_LIVE <- GSEA_reports.DF.plot.nodendrogram.custom.notitle(data = GSEA_reports.DF.custom,top.margin=0,bottom.margin=0,LorB="L")
# Blank
GSEA.timecourse_soil_BLANK <- GSEA_reports.DF.plot.nodendrogram.custom.notitle(data = GSEA_reports.DF.custom,top.margin=0,bottom.margin=0,LorB="B")
# combined
GSEA.timecourse_soil_all <- plot_grid(GSEA.timecourse_soil_LIVE,GSEA.timecourse_soil_BLANK,align="h",labels=c("Blank","Live"))
ggsave2(GSEA.timecourse_soil_all,filename="GSEA.timecourse_soil_all4.png",path=file.path("..","output"),width=20,height=8)

```
